Add internal links to table of contents.
The entries on the first page get inserted in a second iteration because we don't know if all actions fit on one page during the first iteration.
This process could be further improved for some cleaner code by checking if all actions fit on one page first, then adding all texts with links in one iteration.